+++ lib/knossos/algorithm.rb
@@ -1,6 +1,7 @@
 require_relative 'algorithm/aldous_broder'
 require_relative 'algorithm/binary_tree'
 require_relative 'algorithm/recursive_backtracker'
+require_relative 'algorithm/sidewinder'
 require_relative 'algorithm/wilsons'
 
 module Knossos::Algorithm